% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqc.R
\name{aqcApp}
\alias{aqcApp}
\title{Argo QC skeleton}
\usage{
aqcApp(
  data = NULL,
  ID = NULL,
  cycle = NULL,
  age = argoFloats::argoDefaultIndexAge(),
  server = argoFloats::argoDefaultServer(),
  destdir = argoFloats::argoDefaultDestdir(),
  debug = 0
)
}
\arguments{
\item{data}{an oce object (optional). If provided, then \code{ID}, \code{cycle}, \code{age},
\code{server} and \code{destdir} are all ignored.}

\item{ID}{...}

\item{cycle}{...}

\item{age}{numeric value indicating how old a downloaded file
must be (in days), for it to be considered out-of-date.  The
default, \code{argoDefaultIndexAge()}, limits downloads to once per day, as a way
to avoid slowing down a workflow with a download that might take
a minute or so. Note that setting \code{age=0} will force a new
download, regardless of the age of the local file.}

\item{server}{...}

\item{destdir}{...}

\item{debug}{integer value that controls how much information this app prints
to the console as it works.  The default value of 0 leads to a fairly limited
amount of printing, while higher values lead to more information. This information
can be helpful in diagnosing problems or bottlenecks.}
}
\description{
Argo QC skeleton
}
\examples{
# Case 1. pre-loaded data
f <- system.file("extdata/R6903548_029.nc")
if (nchar(f)) {
    d <- oce::read.argo(f)
    aqcApp(d)
}

}
\author{
Dan Kelley
}
