% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceInspect.R
\name{oceInspectApp}
\alias{oceInspectApp}
\title{Inspect an oce CTD-like object}
\usage{
oceInspectApp(objects = NULL)
}
\arguments{
\item{objects}{either a list containing either CTD-like oce objects or a names of
files containing such objects.  For the file-name case, the data are read with
\code{\link[oce:read.oce]{oce::read.oce()}}.}
}
\description{
This is an R shiny app that displays the contents of CTD-like oce objects, i.e.
objects that contain salinity, temperature, pressure, longitude and latitude.
A combination of GUI and keystroke actions (type \verb{?} to list) permit
flagging data points of interest,
and the app can save this information in a CSV file, for later processing.
This can be useful for tasks that benefit from a human eye, including
identification of features such as mixed layers and intrusions, and of
various errors that can result during measurement.
}
\details{
GUI controls are provided for selecting the plotted field, and the app also responds to keystrokes.
A text box shows the current location of the mouse pointer in the (x,y) coordinates of the
graph, as well as the data point that is nearest.  Pressing the \code{0} key (or any other numeric
key, up to \code{9}) saves this information into a buffer along with the name of the input data
and the present plot view.  Points may be removed from this buffer by place the mouse
on them and typing '-'.

The buffer data may be saved to a CSV file (the name of which is patterned on the input
filename). This happens if the context is switched (with the "Name" pulldown menu),
if the \code{w} key is pressed, or if the Quit button is pressed.  If \code{oceInspectApp()} is used
again in the same directory, it will load such files at startup.
}
\examples{
library(oceInspect)
# Handle an argo file and a ctd file.
if (interactive()) {
    f1 <- system.file("extdata/D6903548_029.nc", package="oceInspect")
    f2 <- system.file("extdata/BED0302.cnv", package="oceInspect")
    oceInspectApp(c(f1, f2))
}

}
\author{
Dan Kelley
}
