% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oceInspect.R
\name{oceInspectApp}
\alias{oceInspectApp}
\title{Inspect an oce CTD-like object}
\usage{
oceInspectApp(data = NULL, debug = 0)
}
\arguments{
\item{data}{an oce object (optional). If provided, then \code{ID}, \code{cycle}, \code{age},
\code{server} and \code{destdir} are all ignored.}

\item{debug}{integer value that controls how much information this app prints
to the console as it works.  The default value of 0 leads to a fairly limited
amount of printing, while higher values lead to more information. This information
can be helpful in diagnosing problems or bottlenecks.}
}
\description{
Inspect an oce CTD-like object
}
\examples{
library(oceInspect)
# Example 1: argo file
if (interactive()) {
    f <- system.file("extdata/R6903548_029.nc", package="oceInspect")
    d <- oce::read.oce(f)
    oceInspectApp(d)
}
# Example 2: ctd file
if (interactive()) {
    f <- system.file("extdata/ctd.cnv", package="oceInspect")
    d <- oce::read.oce(f)
    oceInspectApp(d)
}

}
\author{
Dan Kelley
}
